+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## next
 
+- Improved string rendering in the values popup of `signature` view
 - Fixed optional values statistics in `signature` view
 
 ## 1.0.0-beta.89 (04-11-2024)